Separate skynet.go into different files, refactor options
