+++ tests/download_integration_test.go
@@ -29,7 +29,7 @@ func TestDownloadFile(t *testing.T) {
 	//
 	// Match against the full URL, including the skylink.
 	opts := skynet.DefaultDownloadOptions
-	urlpath := strings.TrimRight(opts.PortalDownloadPath, "/") + "/" + skylink
+	urlpath := strings.TrimRight(opts.EndpointPath, "/") + "/" + skylink
 	gock.New(skynet.DefaultPortalURL).
 		Get(urlpath).
 		Reply(200).
@@ -79,7 +79,7 @@ func TestDownloadFileSkykey(t *testing.T) {
 
 	opts := skynet.DefaultDownloadOptions
 	opts.SkykeyName = skykeyName
-	urlpath := strings.TrimRight(opts.PortalDownloadPath, "/") + "/" + skylink
+	urlpath := strings.TrimRight(opts.EndpointPath, "/") + "/" + skylink
 	gock.New(skynet.DefaultPortalURL).
 		Get(urlpath).
 		MatchParam("skykeyname", skykeyName).
@@ -109,7 +109,7 @@ func TestDownloadFileSkykey(t *testing.T) {
 
 	opts = skynet.DefaultDownloadOptions
 	opts.SkykeyID = skykeyID
-	urlpath = strings.TrimRight(opts.PortalDownloadPath, "/") + "/" + skylink
+	urlpath = strings.TrimRight(opts.EndpointPath, "/") + "/" + skylink
 	gock.New(skynet.DefaultPortalURL).
 		Get(urlpath).
 		MatchParam("skykeyid", skykeyID).
