+++ tests/skykeys_integration_test.go
@@ -16,7 +16,7 @@ func TestAddSkykey(t *testing.T) {
 
 	opts := skynet.DefaultAddSkykeyOptions
 	gock.New(skynet.DefaultPortalURL).
-		Post(opts.PortalAddSkykeyPath).
+		Post(opts.EndpointPath).
 		MatchParam("skykey", skykey).
 		Reply(200)
 
@@ -42,7 +42,7 @@ func TestCreateSkykey(t *testing.T) {
 
 	opts := skynet.DefaultCreateSkykeyOptions
 	gock.New(skynet.DefaultPortalURL).
-		Post(opts.PortalCreateSkykeyPath).
+		Post(opts.EndpointPath).
 		MatchParam("name", name).
 		MatchParam("type", skykeyType).
 		Reply(200).
@@ -82,7 +82,7 @@ func TestGetSkykey(t *testing.T) {
 
 	opts := skynet.DefaultGetSkykeyOptions
 	gock.New(skynet.DefaultPortalURL).
-		Get(opts.PortalGetSkykeyPath).
+		Get(opts.EndpointPath).
 		Reply(200).
 		JSON(skynet.Skykey{Skykey: skykey, Name: name, ID: id, Type: skykeyType})
 
@@ -104,7 +104,7 @@ func TestGetSkykey(t *testing.T) {
 	// Get by ID
 
 	gock.New(skynet.DefaultPortalURL).
-		Get(opts.PortalGetSkykeyPath).
+		Get(opts.EndpointPath).
 		Reply(200).
 		JSON(skynet.Skykey{Skykey: skykey, Name: name, ID: id, Type: skykeyType})
 
@@ -147,7 +147,7 @@ func TestListSkykeys(t *testing.T) {
 
 	opts := skynet.DefaultListSkykeysOptions
 	gock.New(skynet.DefaultPortalURL).
-		Get(opts.PortalListSkykeysPath).
+		Get(opts.EndpointPath).
 		Reply(200).
 		JSON(map[string][]skynet.Skykey{"skykeys": response})
